refactor(UserMessage): 클래스 패키지 변경
- view -> util
