+++ src/main/java/util/UserMessage.java
@@ -1,4 +1,4 @@
-package view;
+package util;
 
 public class UserMessage {
     public static final String PLAYER_INPUT_PROMPT = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
@@ -7,6 +7,8 @@ public class UserMessage {
     public static final String LADDER_GENERATE_RESULT_HEADER = "사다리 결과";
     public static final String SEARCH_PLAYER_PROMPT = "결과를 보고 싶은 사람은?";
     public static final String GAME_RESULT_HEADER = "실행 결과";
+
     private UserMessage() {
+        throw new IllegalStateException("Utility class");
     }
 }
